fastapi_backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.4.2
pydantic[email]==2.4.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
psycopg2-binary==2.9.9
numpy==1.24.3
scipy==1.10.1
matplotlib==3.7.2
seaborn==0.12.2
pandas==2.0.3
statsmodels==0.14.0
pywavelets==1.4.1
mne==1.5.1
tensorflow==2.15.0
python-docx==1.0.1 